New folder/backend/requirements.txt
Flask==2.2.3
Flask-RESTful==0.3.9
Flask-JWT-Extended==4.4.4
Flask-Cors==3.0.10
python-dotenv==1.0.0
gunicorn==20.1.0
qrcode==7.4.2
Pillow==9.5.0
pytest==7.3.1
requests==2.28.2